+++ docs/changelog.md
@@ -10,6 +10,7 @@ This is a list of changes, from new to older.
 - Query balances on the marketplace at startup and after each trade. A notification is send such that you always know how much fiat money is left to trade.
 - In the trigger evaluation, triggers are moved to multiple rows if there are more than three of them.
 - Allow trigger evaluation on a chosen subrange of dates.
+- Present trigger evaluation summary as a table.
 
 ### Version 3.3.0